pyproject.toml
[project]
name = "choose-adventure"
version = "0.1.0"
description = "A CLI choose-your-own-adventure impact clarity tool for EA"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "rich>=13.0.0",
    "networkx>=3.0",
    "pyyaml>=6.0",
]

[project.scripts]
adventure = "adventure.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/adventure"]
